refactor(user_address): Migrate address_entity to shared_models and update imports
