+++ lib/core/utils/routes/routes.dart
@@ -1,4 +1,3 @@
-import 'package:ecommerce_flower_app/features/saved_addresses/domain/entity/address_entity.dart';
 import 'package:flutter/material.dart';
 
 import '../../../features/about_us/presentation/view/about_us_screen.dart';
@@ -17,6 +16,7 @@ import '../../../features/product_details/presentation/view/screens/product_deta
 import '../../../features/most_selling_products/presentation/view/screens/best_sellers_screen.dart';
 import '../../../features/saved_addresses/presentation/view/saved_address_screen.dart';
 import '../../../features/terms_and_conditions/presentation/view/terms_and_conditions_screen.dart';
+import '../shared_models/address_entity.dart';
 import '../shared_models/product_entity.dart';
 
 class AppRoutes {
