+++ lib/features/address_details/presentation/view_model/address_details_cubit.dart
@@ -2,13 +2,13 @@ import 'package:ecommerce_flower_app/core/base/base_state.dart';
 import 'package:ecommerce_flower_app/core/utils/datasource_excution/api_result.dart';
 import 'package:ecommerce_flower_app/core/utils/validator/validator.dart';
 import 'package:ecommerce_flower_app/features/address_details/domain/use_cases/get_cities_use_case.dart';
-import 'package:ecommerce_flower_app/features/saved_addresses/domain/entity/address_entity.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:geocoding/geocoding.dart';
 import 'package:injectable/injectable.dart';
 import 'package:latlong2/latlong.dart';
+import '../../../../core/utils/shared_models/address_entity.dart';
 import '../../domain/entities/address_details_request_entity.dart';
 import '../../domain/use_cases/add_adderss_use_case.dart';
 import '../../domain/use_cases/get_states_use_case.dart';
