--- lib/features/saved_addresses/domain/entity/address_entity.dart
@@ -1,19 +1,0 @@
-class AddressEntity {
-  final String? id;
-  final String? street;
-  final String? city;
-  final String? lat;
-  final String? long;
-  final String? phone;
-  final String? username;
-
-  AddressEntity({
-    this.id,
-    this.street,
-    this.city,
-    this.lat,
-    this.long,
-    this.phone,
-    this.username,
-  });
-}
