+++ lib/features/saved_addresses/domain/use_case/delete_user_address_use_case.dart
@@ -1,7 +1,7 @@
 import 'package:injectable/injectable.dart';
 
 import '../../../../core/utils/datasource_excution/api_result.dart';
-import '../entity/address_entity.dart';
+import '../../../../core/utils/shared_models/address_entity.dart';
 import '../repo/saved_addresses_repo.dart';
 
 @injectable
